% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_to_txt.R
\name{dir_to_txt}
\alias{dir_to_txt}
\title{Zip and encode a directory to csv file}
\usage{
dir_to_txt(
  dir_path,
  txt_file_path,
  encoding_scheme = c("base64", "xxd"),
  exclude_dot_R = TRUE,
  return_zip_output = FALSE
)
}
\arguments{
\item{return_zip_output}{}
}
\description{
Zip and encode a directory to csv file
}
