% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Zip, encode a directory and then publish it to download location}
\usage{
export(dir_path = ".")
}
\arguments{
\item{dir_path}{}
}
\description{
Zip, encode a directory and then publish it to download location
}
\examples{
export()
}
