% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_csv.R
\name{import_from_csv}
\alias{import_from_csv}
\title{Import encoded directory from csv}
\usage{
import_from_csv(
  csv_path,
  dest_path = ".",
  project_name = NULL,
  open_project = TRUE
)
}
\arguments{
\item{csv_path}{path of csv file}

\item{dest_path}{path of destination directory}

\item{open_project}{logical.}
}
\value{
nothing
}
\description{
Import encoded directory from csv
}
\examples{
import_from_csv (csv_path, dest_path)
}
