% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_orauser.R
\name{import_from_orauser}
\alias{import_from_orauser}
\title{Import directory form base64-encoded Oracle table}
\usage{
import_from_orauser(TABLE_NAME, dest_path = ".", open_project = TRUE)
}
\arguments{
\item{TABLE_NAME}{oracle table name with one base64-encoded column}

\item{dest_path}{path of destination directory}

\item{open_project}{logical.}
}
\value{
nothing
}
\description{
Import directory form base64-encoded Oracle table
}
\examples{
import_from_orauser(PACKAGE_TABLE_NAME)
}
